requirements.txt
# Core Data & Science
numpy
pandas
matplotlib
seaborn
scipy
scikit-learn
tqdm
pyyaml
pyarrow

# Geospatial & Network
osmnx
geopandas

# Simulation & RL
traci
sumolib
gymnasium

# Intelligence (PyTorch with Metal Support)
torch
torchvision
torchaudio

# Workflow & Tooling
ruff
black
ipykernel